Legger Vedtaksperiode på OpphørLæremidlerRequest
