+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/læremidler/dto/InnvilgelseLæremidlerDto.kt
@@ -15,11 +15,3 @@ data class InnvilgelseLæremidlerRequest(
     val vedtaksperioder: List<VedtaksperiodeDto>,
 ) : VedtakLæremidlerRequest, VedtakLæremidlerDto(TypeVedtak.INNVILGELSE)
 
-data class VedtaksperiodeDto(
-    val fom: LocalDate,
-    val tom: LocalDate,
-)
-
-fun List<Vedtaksperiode>.tilDto() = this.map { VedtaksperiodeDto(fom = it.fom, tom = it.tom) }
-
-fun List<VedtaksperiodeDto>.tilDomene() = this.map { Vedtaksperiode(fom = it.fom, tom = it.tom) }
