+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/læremidler/dto/OpphørLæremidlerDto.kt
@@ -10,6 +10,7 @@ data class OpphørLæremidlerResponse(
 ) : VedtakLæremidlerResponse, VedtakLæremidlerDto(TypeVedtak.OPPHØR)
 
 data class OpphørLæremidlerRequest(
+    val vedtaksperioder: List<VedtaksperiodeDto>,
     val årsakerOpphør: List<ÅrsakOpphør>,
     val begrunnelse: String,
 ) : VedtakLæremidlerRequest, VedtakLæremidlerDto(TypeVedtak.OPPHØR)
