+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/læremidler/dto/VedtakLæremidlerDto.kt
@@ -5,6 +5,8 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo
 import no.nav.tilleggsstonader.sak.vedtak.TypeVedtak
 import no.nav.tilleggsstonader.sak.vedtak.dto.VedtakRequest
 import no.nav.tilleggsstonader.sak.vedtak.dto.VedtakResponse
+import no.nav.tilleggsstonader.sak.vedtak.læremidler.domain.Vedtaksperiode
+import java.time.LocalDate
 
 @JsonTypeInfo(
     use = JsonTypeInfo.Id.NAME,
@@ -22,3 +24,12 @@ sealed class VedtakLæremidlerDto(open val type: TypeVedtak)
 
 sealed interface VedtakLæremidlerRequest : VedtakRequest
 sealed interface VedtakLæremidlerResponse : VedtakResponse
+
+data class VedtaksperiodeDto(
+    val fom: LocalDate,
+    val tom: LocalDate,
+)
+
+fun List<Vedtaksperiode>.tilDto() = this.map { VedtaksperiodeDto(fom = it.fom, tom = it.tom) }
+
+fun List<VedtaksperiodeDto>.tilDomene() = this.map { Vedtaksperiode(fom = it.fom, tom = it.tom) }
